+++ src/requests/event.ts
@@ -2,9 +2,9 @@
 
 import {YcmLocation, YcmRange, HandleRequestError} from './utils'
 import {YcmServer} from '../server'
-import {Diagnostic, DiagnosticSeverity, TextDocument, DiagnosticRelatedInformation, Location} from 'vscode'
+import {Diagnostic, DiagnosticSeverity, TextDocument, DiagnosticRelatedInformation, Location, Uri} from 'vscode'
 import { YcmSimpleRequest } from './simpleRequest';
-import { YcmExtendedDiagnosticRequest } from './extendedDiagnostic';
+import { YcmExtendedDiagnosticRequest, YcmExtendedDiagnosticResponse, YcmExtendedDiagnostic } from './extendedDiagnostic';
 import { Log } from '../utils';
 
 type YcmEvent = 
@@ -21,23 +21,8 @@ export class YcmEventNotification extends YcmSimpleRequest
 
 	async Send(server: YcmServer): Promise<YcmDiagnosticsResponse>
 	{
-		try
-		{
-			let pResRaw = server.SendData('/event_notification', this)
-			let res = await pResRaw
-			return new YcmDiagnosticsResponse(res)
-		}
-		catch(err)
-		{
-			if(await HandleRequestError(err))
-			{
-				return this.Send(server)
-			}
-			else
-			{
-				//TODO: return empty response
-			}
-		}
+		let res = await super.Send(server, '/event_notification')
+		return new YcmDiagnosticsResponse(res)
 	}
 
 }
@@ -49,7 +34,8 @@ export class YcmDiagnosticData
 	location_extent: YcmRange
 	text: string
 	kind: "WARNING" | "ERROR"
-	extendedDiagnostic
+	private extendedDiags: YcmExtendedDiagnostic[]
+	private pExtendedDiags: Promise<YcmExtendedDiagnosticResponse>
 
 	constructor(diagnostic: any)
 	{
@@ -58,47 +44,70 @@ export class YcmDiagnosticData
 		this.location_extent = YcmRange.FromSimpleObject(diagnostic.location_extent)
 		this.text = diagnostic.text
 		this.kind = diagnostic.kind
+		this.pExtendedDiags = this.RequestExtendedDiag()
+		this.extendedDiags = null
 	}
 
-	public async ToVscodeDiagnostic(contextDoc: TextDocument): Promise<Diagnostic>
+	private async RequestExtendedDiag(): Promise<YcmExtendedDiagnosticResponse>
+	{
+		let detailReq = new YcmExtendedDiagnosticRequest(this)
+		return detailReq.Send(await YcmServer.GetInstance())
+	}
+
+	public async ResolveExtendedDiags(): Promise<void>
+	{
+		if(this.pExtendedDiags !== null)
+		{
+			try
+			{
+				this.extendedDiags = (await this.pExtendedDiags).extendedDiags;
+			}
+			catch(e)
+			{
+				Log.Debug("Resolving extended diags failed: ", e);
+				this.extendedDiags = []
+			}
+			this.pExtendedDiags = null
+		}
+	}
+
+	public SetExtendedDiags(diags: YcmExtendedDiagnostic[])
+	{
+		this.pExtendedDiags = null
+		this.extendedDiags = diags
+	}
+
+	public GetExtendedDiags(): YcmExtendedDiagnostic[]
+	{
+		if(this.extendedDiags === null)
+		{
+			throw "Attempted to get extended diags before resolving them"
+		}
+		return this.extendedDiags
+	}
+
+	public async ToVscodeDiagnostic(): Promise<Diagnostic>
 	{
 		let kind = this.kind === "WARNING" ? DiagnosticSeverity.Warning : DiagnosticSeverity.Error
 		let result: Diagnostic
-		let detailReq = new YcmExtendedDiagnosticRequest(this);
-		let pExtendedDiag = detailReq.Send(await YcmServer.GetInstance());
-		if(this.location.filepath == contextDoc.fileName)
-		{
-			result = new Diagnostic(
-				await this.location_extent.ToVscodeRange(), 
-				this.text,
-				kind
-			)
-		}
-		//TODO: check if the location is in the workspace, an place it in that document if it is
-		else
-		{
-			let realLoc = await this.location.GetVscodeLoc()
-			result = new Diagnostic(
-				contextDoc.lineAt(0).range,
-				//even most programmers expect 1-based indexing in file coordinates
-				//still translate to VScode pos in order to get characters instead of byte offsets
-				`${this.text} at ${this.location.filepath}:${realLoc.pos.line+1}:${realLoc.pos.character+1}`,
-				kind
-			)
-		}
+		result = new Diagnostic(
+			await this.location_extent.ToVscodeRange(), 
+			this.text,
+			kind
+		)
 		try
 		{
 			//the diag must be complete when assigned to the collection.
 			//assigning the related info afterwards doesn't work
 			//OTOH, the diags are awaited in parallel, so ¯\_(ツ)_/¯
-			let extendedDiag = await pExtendedDiag
-			let pDiags = extendedDiag.extendedDiags.map(async (extendedDiag) => {
+			
+			let pDiags = this.extendedDiags.map(async (extendedDiag) => {
 				return new DiagnosticRelatedInformation(
 					new Location(
-						contextDoc.uri,
-						(await this.location.GetVscodeLoc()).pos
+						Uri.file(extendedDiag.location.filepath),
+						(await extendedDiag.location.GetVscodeLoc()).pos
 					),
-					extendedDiag
+					extendedDiag.text
 				)
 			})
 			result.relatedInformation = await Promise.all(pDiags)
@@ -107,7 +116,7 @@ export class YcmDiagnosticData
 		{
 			//Exception. This is just extended info, so probably not important. 
 			//Log as unimportant and otherwise ignore
-			Log.Debug("Getting extended diags failed: ", e);
+			Log.Debug("Creating extended diags failed: ", e);
 		}
 		return result
 	}