+++ CHANGELOG.md
@@ -5,6 +5,11 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [v0.2.2]
+
+- Improve diagnostics so that every diagnostic is displayed in 
+	the file it is reported in, fix some problems in communication with the server.
+
 ## [v0.2.0]
 
 - Add extended diagnostic information